+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Add method to replace filter and specify operator on `listing` - [ripe-pulse/#240](https://github.com/ripe-tech/ripe-pulse/issues/240)
+* Add method to remove filters on `listing` - [ripe-pulse/#240](https://github.com/ripe-tech/ripe-pulse/issues/240)
 
 ### Changed